--- dev/sbt-plugin/src/main/scala-sbt-1.0/com/lightbend/lagom/sbt/LagomImportCompat.scala
@@ -1,11 +1,0 @@
-/*
- * Copyright (C) 2016-2019 Lightbend Inc. <https://www.lightbend.com>
- */
-
-package com.lightbend.lagom.sbt
-
-import sbt.ForkOptions
-
-trait LagomImportCompat {
-  def getForkOptions(options: Vector[String]): ForkOptions = ForkOptions().withRunJVMOptions(options)
-}
